v1.1

+++ routers/serviceRouter.js
@@ -1,25 +1,40 @@
 const router = require("express").Router();
 const Service = require("../models/serviceModel");
+const auth = require("../middleware/auth");
 
 
-router.get("/", async (req, res) => {
-    const services = await Service.find();
-    res.json(services);
+router.get("/", auth, async (req, res) => {
+
+    try{
+        
+        const services = await Service.find({user: req.user});
+        res.json(services);
+
+    } catch (err) {
+        res.status(500).send();
+    }
+   
 });
 
-router.post("/", async (req, res) => {
+router.post("/", auth, async (req, res) => {
     try{
 
-    const {name, service, link} = req.body;
+    const {name, location, service, price, link} = req.body;
 
     //validation
 
-    if(!service || !link) {
-        return res.status(400).json({msg: "Please provide the service you would like to add with relevant link"})
+    if(!service || !location || !link || !name || !price) {
+        return res.status(400).json({msg: "Please provide all fields"})
     }
     
     const newService = new Service({
-        name, service, link
+        name, 
+        location,
+        service, 
+        price,
+        link,
+        user: req.user
+
     })
 
     const savedService = await newService.save();
@@ -30,14 +45,16 @@ router.post("/", async (req, res) => {
 }
 });
 
-router.put("/:id", async (req, res) => {
+router.put("/:id", auth, async (req, res) => {
     try{
-        const { name, service, link} = req.body;
+
+        
+        const { name, location, service, price, link} = req.body;
         const serviceId = req.params.id;
 
         //validation
-        if(!service || !link) {
-            return res.status(400).json({msg: "Please provide the service you would like to add with relevant link"})
+        if(!service || !location || !link || !name || !price) {
+            return res.status(400).json({msg: "Please fill in all fields"})
         };
 
         if(!serviceId)
@@ -47,8 +64,14 @@ router.put("/:id", async (req, res) => {
         if(!originalService)
             return res.status(400).json({msg: "Please contact the developer"});
 
+
+        if(originalService.user.toString() !== req.user)
+            return res.status(401).json({msg: "unauthorised"});
+
        originalService.name = name;
+       originalService.location = name;
        originalService.service = service;
+       originalService.price = price;
        originalService.link = link;
 
        const savedService = await originalService.save();
@@ -60,7 +83,7 @@ router.put("/:id", async (req, res) => {
 })
 
 
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", auth, async (req, res) => {
     try{
         const serviceId = req.params.id;
 
@@ -71,6 +94,9 @@ router.delete("/:id", async (req, res) => {
         const existingService = await Service.findById(serviceId);
         if(!existingService)
             return res.status(400).json({msg: "Please contact the developer"});
+
+        if(existingService.user.toString() !== req.user)
+            return res.status(401).json({msg: "unauthorised"});
 
         await existingService.delete();
